+++ src/components/common/Pagination/index.tsx
@@ -11,8 +11,6 @@ interface PaginationProps {
   setState: Dispatch<SetStateAction<number>>;
 }
 
-const PAGE_OFFSET = 10;
-
 const Pagination = (props: PaginationProps) => {
   const { currentPage, totalItems, itemsPerPage, setState } = props;
 
@@ -53,4 +51,4 @@ const Pagination = (props: PaginationProps) => {
   );
 };
 
-export { Pagination, PAGE_OFFSET };
+export { Pagination };
