+++ src/components/notice/table/TableContent/index.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 
-import { SIMMAJOR } from '@/constants/major';
 import { NoticeResponse } from '@/types/notice';
+import { YYMMDD } from '@/utils/date';
 
 interface TableContentProps {
   contents: NoticeResponse[];
@@ -11,26 +11,13 @@ interface TableContentProps {
 const TableContent = (props: TableContentProps) => {
   const { contents, handleContent } = props;
 
-  const formatDate = (dateString: string) => {
-    const date = new Date(dateString);
-    const year = date.getFullYear().toString().slice(-2);
-    const month = (date.getMonth() + 1).toString().padStart(2, '0');
-    const day = date.getDate().toString().padStart(2, '0');
-    return `${year}${month}${day}`;
-  };
-
-  const getMajorName = (major: number) => {
-    return SIMMAJOR[major] || '';
-  };
-
   return (
     <tbody>
       {contents.map((notice: NoticeResponse) => (
         <Styled.Row key={notice.id} onClick={() => handleContent(notice.id)}>
           <Styled.Item>{notice.id}</Styled.Item>
-          <Styled.Item>{getMajorName(notice.major)}</Styled.Item>
           <Styled.Item>{notice.title}</Styled.Item>
-          <Styled.Item>{formatDate(notice.created_at)}</Styled.Item>
+          <Styled.Item>{YYMMDD(notice.created_at)}</Styled.Item>
         </Styled.Row>
       ))}
     </tbody>
