+++ src/constants/major.ts
@@ -34,40 +34,3 @@ export const MAJOR: { [key: string]: number } = {
   국제학부: 32,
   LD학부: 33,
 };
-
-export const SIMMAJOR: { [key: number]: string } = {
-  0: '학과',
-  1: '프어',
-  2: '독어',
-  3: '노어',
-  4: '서어',
-  5: '이어',
-  6: '포어',
-  7: '네어',
-  8: '스칸',
-  9: '마인',
-  10: '아랍',
-  11: '태국',
-  12: '베어',
-  13: '인도',
-  14: '터키',
-  15: '페이',
-  16: '몽골',
-  17: 'ELLT',
-  18: '영문',
-  19: 'EICC',
-  20: '중국',
-  21: '일본',
-  22: '정외',
-  23: '행정',
-  24: '미컴',
-  25: '영교',
-  26: '한교',
-  27: '프교',
-  28: '독교',
-  29: '중교',
-  30: '상경',
-  31: '경영',
-  32: '국제',
-  33: 'LD',
-};
