+++ src/constants/pageoffset.ts
@@ -1,0 +1,1 @@
+export const PAGE_OFFSET = 10;
