+++ src/query/notice.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQuery, useQueryClient } from 'react-query';
 
-import { delNotice, getMajorNotice, getNotice, postNotice } from '@/api/notice';
+import { delNotice, getNotice, postNotice } from '@/api/notice';
 import { MAJOR } from '@/constants/major';
 import useToast from '@/hooks/useToast';
 import { useFetchMe } from '@/query/user';
@@ -11,28 +11,13 @@ const QUERY_KEY = {
   id: 'id',
 };
 
-export const useFetchNotice = () => {
+export const useFetchNotice = (major: number) => {
   const { createToastMessage } = useToast();
-  const { data: notices, isLoading } = useQuery(QUERY_KEY.notice, getNotice, {
+  const { data: notices, isLoading } = useQuery([QUERY_KEY.notice, major], () => getNotice(major), {
     onError: () => {
       createToastMessage('다시 시도해주세요.', 'error');
     },
   });
-
-  return { data: notices, isLoading };
-};
-
-export const useFetchMajorNotice = (major: number) => {
-  const { createToastMessage } = useToast();
-  const { data: notices, isLoading } = useQuery(
-    [QUERY_KEY.notice, major],
-    () => getMajorNotice(major),
-    {
-      onError: () => {
-        createToastMessage('다시 시도해주세요.', 'error');
-      },
-    }
-  );
   return { data: notices, isLoading };
 };
 
@@ -62,11 +47,8 @@ export const useCreateNoticeMutation = () => {
 export const useDeleteNoticeMutation = () => {
   const { createToastMessage } = useToast();
   const queryClient = useQueryClient();
-  const deleteNotice = (id: number) => {
-    return delNotice(id);
-  };
 
-  const mutation = useMutation(deleteNotice, {
+  const mutation = useMutation(delNotice, {
     onSuccess: () => {
       createToastMessage('공지사항 삭제 완료!', 'success');
       queryClient.invalidateQueries(QUERY_KEY.notice);
