+++ src/utils/date.ts
@@ -7,3 +7,12 @@ export const formatDate = (date: Date) => {
 
   return `${month}월 ${day}일 ${weekday}`;
 };
+
+export const YYMMDD = (datestring: string) => {
+  const date = new Date(datestring);
+  const year = date.getFullYear().toString().slice(-2);
+  const month = (date.getMonth() + 1).toString().padStart(2, '0');
+  const day = date.getDate().toString().padStart(2, '0');
+
+  return `${year}${month}${day}`;
+};
